# backend/paper-service/src/main/resources/openapi/papers/paper_api.yaml
openapi: 3.0.3

info:
  title: Requests for Service Papers
  version: 1.0.0
  description: OpenApi Specification for Service Papers

servers:
  - url: /
    description: Default Server URL

paths:

  /papers/upload/{username}:
    post:
      tags:
        - Papers
      summary: Метод загрузки статьи пользователя на сервер
      description: Метод загрузки статьи пользователя на сервер
      operationId: upload
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                paper:
                  type: string
                  format: binary
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/upload_paper_response.json'
        400:
          description: Incorrect request format
        404:
          description: User not found
        500:
          description: Internal server error

  /papers/check/{username}/{paperId}:
    post:
      tags:
        - Papers
      summary: Метод проверки статьи на плагиат
      description: Метод проверки статьи на плагиат
      operationId: check
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/check_paper_response.json'
        400:
          description: Incorrect request format
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/checks/{username}:
    get:
      tags:
        - Papers
      summary: Метод получения проверок статей на плагиат
      description: Метод получения проверок статей на плагиат
      operationId: getChecks
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type : array
                items:
                  $ref: 'model/status_paper_response.json'
        400:
          description: Incorrect request format
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/get/{username}:
    get:
      tags:
        - Papers
      summary: Метод получения статей пользователя
      description: Метод получения статей пользователя
      operationId: get
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/get_papers_response.json'
        400:
          description: Incorrect request format
        403:
          description: Permission denied
        409:
          description: Invalid items in request
        429:
          description: To many requests
        500:
          description: Internal server error


  /papers/delete/{username}/{paperId}:
    delete:
      tags:
        - Papers
      summary: Метод удаления статьи пользователя из системы
      description: Метод удаления статьи пользователя из системы
      operationId: delete
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        400:
          description: Incorrect request format
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/status/{username}/{paperId}:
    get:
      tags:
        - Papers
      summary: Метод получения информации о текущей проверке статьи
      description: Метод получения информации о текущей проверке статьи
      operationId: status
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/status_paper_response.json'
        400:
          description: Incorrect request format
        403:
          description: Auth error
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/group/upload/{groupId}:
    post:
      tags:
        - PapersGroup
      summary: Метод загрузки статьи пользователя на сервер
      description: Метод загрузки статьи пользователя на сервер
      operationId: uploadGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                paper:
                  type: string
                  format: binary
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/upload_paper_response.json'
        400:
          description: Incorrect request format
        404:
          description: User not found
        500:
          description: Internal server error

  /papers/group/check/{groupId}/{paperId}:
    post:
      tags:
        - PapersGroup
      summary: Метод проверки статьи на плагиат
      description: Метод проверки статьи на плагиат
      operationId: checkGroup
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/check_paper_response.json'
        400:
          description: Incorrect request format
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/group/checks/{groupId}:
    get:
      tags:
        - PapersGroup
      summary: Метод получения проверок статей на плагиат
      description: Метод получения проверок статей на плагиат
      operationId: getChecksGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'model/status_paper_response.json'
        400:
          description: Incorrect request format
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/group/get/{groupId}:
    get:
      tags:
        - PapersGroup
      summary: Метод получения статей пользователя
      description: Метод получения статей пользователя
      operationId: getGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/get_papers_response.json'
        400:
          description: Incorrect request format
        403:
          description: Permission denied
        409:
          description: Invalid items in request
        429:
          description: To many requests
        500:
          description: Internal server error


  /papers/group/delete/{groupId}/{paperId}:
    delete:
      tags:
        - PapersGroup
      summary: Метод удаления статьи пользователя из системы
      description: Метод удаления статьи пользователя из системы
      operationId: deleteGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        400:
          description: Incorrect request format
        429:
          description: To many requests
        500:
          description: Internal server error

  /papers/group/status/{groupId}/{paperId}:
    get:
      tags:
        - PapersGroup
      summary: Метод получения информации о текущей проверке статьи
      description: Метод получения информации о текущей проверке статьи
      operationId: statusGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: 'model/status_paper_response.json'
        400:
          description: Incorrect request format
        403:
          description: Auth error
        429:
          description: To many requests
        500:
          description: Internal server error

tags:
  - name: Papers
    description: "API для работы с сервисом Papers"